requirements.txt
annotated-types==0.7.0
anyio==4.9.0
ariadne==0.26.2
bcrypt==4.3.0
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
filelock==3.24.0
graphql-core==3.2.5
h11==0.16.0
httptools==0.7.1
identify==2.6.16
idna==3.11
motor==3.7.0
nodeenv==1.10.0
platformdirs==4.9.0
pre_commit==4.5.1
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.16.0
pyrefly==0.24.2
python-dotenv==1.2.1
PyYAML==6.0.3
ruff==0.15.1
sniffio==1.3.1
starlette==0.41.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.30.6
uvloop==0.22.1
virtualenv==20.36.1
watchfiles==1.1.1
websockets==15.0.1
